Fix property card display: replace hardcoded 'Farmplot' with real propertyType
 PropertyCard now shows real property types: VILLA, HOLIDAY_HOME, APARTMENT, etc.
 Improved price parsing and formatting
 Added proper spacing in display

Cards now show: '1,200 sqft | VILLA | 25 acres' instead of '1200 sqft | Farmplot | 12 Acres'

+++ src/app/api/properties/route.ts
@@ -202,7 +202,7 @@ export async function GET(request: NextRequest) {
       try {
         return {
           ...property,
-          price: parseFloat(property.price) || 0,
+          price: parseInt(property.price) || 0,
           images: parseArray(property.images),
           amenities: parseArray(property.amenities),
           features: parseArray(property.features),
@@ -226,7 +226,7 @@ export async function GET(request: NextRequest) {
         // Return the property with basic fields only if parsing fails
         return {
           ...property,
-          price: parseFloat(property.price) || 0,
+          price: parseInt(property.price) || 0,
           images: [],
           amenities: [],
           features: [],
